+++ packages/desmos-std/src/mock.rs
@@ -32,142 +32,23 @@ pub fn mock_dependencies_with_custom_querier(
 #[cfg(test)]
 mod tests {
     use crate::{
+        mock::mock_dependencies_with_custom_querier,
         profiles::{
-            mock::{MockProfilesQuerier, MockProfilesQueries},
-            models_query::{
-                QueryApplicationLinkByClientIDResponse, QueryApplicationLinksResponse,
-                QueryBlocksResponse, QueryChainLinksResponse,
-                QueryIncomingDtagTransferRequestResponse, QueryProfileResponse,
-                QueryRelationshipsResponse, QueryUserApplicationLinkResponse,
-                QueryUserChainLinkResponse,
-            },
-            query_router::ProfilesQuery,
+            mock::MockProfilesQueries, models_query::QueryProfileResponse, querier::ProfilesQuerier,
         },
-        query::DesmosQuery,
     };
-    use cosmwasm_std::{to_binary, Addr};
+    use cosmwasm_std::Addr;
+    use std::ops::Deref;
 
     #[test]
-    fn test_query_profile() {
-        let query = ProfilesQuery::Profile {
-            user: Addr::unchecked(""),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryProfileResponse {
+    fn test_profiles_querier_mock() {
+        let owned_deps = mock_dependencies_with_custom_querier(&[]);
+        let deps = owned_deps.as_ref();
+        let querier = ProfilesQuerier::new(deps.querier.deref());
+        let response = querier.query_profile(Addr::unchecked("")).unwrap();
+        let expected = QueryProfileResponse {
             profile: MockProfilesQueries::get_mock_profile(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_incoming_dtag_transfer_requests() {
-        let query = ProfilesQuery::IncomingDtagTransferRequests {
-            receiver: Addr::unchecked(""),
-            pagination: Default::default(),
         };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryIncomingDtagTransferRequestResponse {
-            requests: vec![MockProfilesQueries::get_mock_dtag_transfer_request()],
-            pagination: Default::default(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_relationships() {
-        let query = ProfilesQuery::Relationships {
-            user: Addr::unchecked(""),
-            subspace_id: 0,
-            pagination: Default::default(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryRelationshipsResponse {
-            relationships: vec![MockProfilesQueries::get_mock_relationship()],
-            pagination: Default::default(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_blocks() {
-        let query = ProfilesQuery::Blocks {
-            user: Addr::unchecked(""),
-            subspace_id: 0,
-            pagination: Default::default(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryBlocksResponse {
-            blocks: vec![MockProfilesQueries::get_mock_user_block()],
-            pagination: Default::default(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_chain_links() {
-        let query = ProfilesQuery::ChainLinks {
-            user: Addr::unchecked(""),
-            pagination: Default::default(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryChainLinksResponse {
-            links: vec![MockProfilesQueries::get_mock_chain_link()],
-            pagination: Default::default(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_user_chain_link() {
-        let query = ProfilesQuery::UserChainLink {
-            user: Addr::unchecked(""),
-            chain_name: "cosmos".to_string(),
-            target: "".to_string(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryUserChainLinkResponse {
-            link: MockProfilesQueries::get_mock_chain_link(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_app_links() {
-        let query = ProfilesQuery::AppLinks {
-            user: Addr::unchecked(""),
-            pagination: Default::default(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryApplicationLinksResponse {
-            links: vec![MockProfilesQueries::get_mock_application_link()],
-            pagination: Default::default(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_user_app_links() {
-        let query = ProfilesQuery::UserAppLinks {
-            user: Addr::unchecked(""),
-            application: "".to_string(),
-            username: "".to_string(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryUserApplicationLinkResponse {
-            link: MockProfilesQueries::get_mock_application_link(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
-    }
-
-    #[test]
-    fn test_query_application_link_by_chain_id() {
-        let query = ProfilesQuery::ApplicationLinkByChainID {
-            client_id: "".to_string(),
-        };
-        let response = MockProfilesQuerier::query(&DesmosQuery::from(query));
-        let expected = to_binary(&QueryApplicationLinkByClientIDResponse {
-            link: MockProfilesQueries::get_mock_application_link(),
-        });
-        assert_eq!(response.into_result().ok(), expected.ok())
+        assert_eq!(expected, response)
     }
 }
