+++ packages/desmos-std/src/query.rs
@@ -27,3 +27,25 @@ impl From<ProfilesQuery> for DesmosQuery {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        profiles::query_router::ProfilesQuery,
+        query::{DesmosQuery, DesmosQueryRoute},
+        types::DesmosRoute,
+    };
+    use cosmwasm_std::Addr;
+
+    #[test]
+    fn test_from_profiles_msg() {
+        let query = ProfilesQuery::Profile {
+            user: Addr::unchecked("cosmos18xnmlzqrqr6zt526pnczxe65zk3f4xgmndpxn2"),
+        };
+        let expected = DesmosQuery {
+            route: DesmosRoute::Profiles,
+            query_data: DesmosQueryRoute::Profiles(query.clone()),
+        };
+        assert_eq!(expected, DesmosQuery::from(query));
+    }
+}
